Refactor estimated duration fields to use `double` instead of `int`
Updated various classes and database schema to change the data type of `estimatedDuration` from `int` to `double`. This change allows for more precise estimation of durations, accommodating fractional hours. Adjusted related methods and database changelogs to reflect this modification, ensuring consistency across the codebase and database.

+++ api/src/main/java/com/grash/dto/WorkOrderBasePatchDTO.java
@@ -14,7 +14,7 @@ public class WorkOrderBasePatchDTO {
 
     private Date dueDate;
     private Priority priority = Priority.NONE;
-    private int estimatedDuration;
+    private double estimatedDuration;
     private Date estimatedStartDate;
     private String description;
     private String title;
