+++ api/src/main/java/com/grash/dto/analytics/workOrders/WOStatsByPriority.java
@@ -21,6 +21,6 @@ public class WOStatsByPriority {
     @AllArgsConstructor
     public static class BasicStats {
         private int count;
-        private int estimatedHours;
+        private double estimatedHours;
     }
 }
